feat: introduce configuration for UPM
Create folder com.xreducation.questvisionstreamplugin as main folder for UPM
this removes the previous .androidlib configuration and separates the library
into a folder for Android Studio. From there we generate .aar and palce into the UPM folder.
Adding extra scripts to import the core scripts into the project once imported
This is a checkpoint for the v1.0.0